requirements.txt
Django==1.5.4
konira
sure
djangorestframework
django-guardian
gunicorn
dj-database-url
dj-static
